% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostwriteR.R
\name{ghostwriter_graph}
\alias{ghostwriter_graph}
\title{Build a DiagrammeR graph from extracted edges}
\usage{
ghostwriter_graph(x)
}
\arguments{
\item{x}{A parsed ghostwriteR object or an edge data frame from \code{ghostwriter_edges()}.}
}
\value{
A DiagrammeR grViz object.
}
\description{
Build a DiagrammeR graph from extracted edges
}
